pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mike_x_webhook_server"
version = "0.1.0"
description = "你的项目描述"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
authors = [
    { name = "Morboz", email = "xx_695513639@126.com" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "flask",
    "requests",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "isort>=5.0",
    "mypy>=1.0",
]

[project.urls]
Homepage = "https://github.com/Morboz/mike-x-form-webhook-server"
Documentation = "https://your_package.readthedocs.io/"
Repository = "https://github.com/Morboz/mike-x-form-webhook-server"

[tool.hatch.build.targets.wheel]
packages = ["src/mike_x_webhook_server"]

[tool.pytest.ini_options]
addopts = "-ra -q"
testpaths = [
    "tests",
]

[tool.black]
line-length = 88
target-version = ["py38"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3